Meant to say that coders, NOT users, can adjust combination of mouse buttons to set for drag scrolling.
Fixed UI issue when clicking any of the edit buttons for users BEFORE
the user with the leader button ON.

Scrollbars are now tracked and will retain their values when switching
from view mode to edit mode and vice versa. Will even retain after
closing and reopening the UI.

Currently writing code to track main window when it's maximized or not.